+++ src/frontend/src/components/EditProfile.svelte
@@ -7,7 +7,6 @@
   import { ROUTES } from "../router/routes";
   import { Kind } from "nostr-tools";
 
-
   let loading = false;
   let profile : Profile = JSON.parse(JSON.stringify(nostric_user.get_profile()));
 
@@ -16,12 +15,12 @@
     let response : Result = await actor.updateProfile(profile);
     if ("ok" in response) {
       nostric_user.set_profile(response.ok);
-      // Publish updated profile to the Nostr relay as well
-      let content = JSON.stringify({
+      // publish updated profile to the Nostr relay as well
+      let content = {
         "username": profile.username,
         "about": profile.about,
         "picture": profile.avatar_url
-      });
+      };
       let event = nostr_service.create_event(content, Kind.Metadata);
       await nostr_service.publish_event(event);
       await navigateTo(ROUTES.POSTS);
