+++ src/frontend/src/components/Spinner.svelte
